Add a consistent rendering protocol.
This change provides a consistent API to render a htpy object as HTML or
iterate over it.

This commit introduces stream_chunks() which is identical with
__iter__() but with a better name.

With the introduction of Fragment, this commit makes render_node and
iter_node redundant and they will be deprecated in another commit.

This commit removes all internal usages of render_node and iter_node so
they can be deprecated.

More info: https://github.com/pelme/htpy/pull/86#discussion_r1986098069