Split htpy into multiple private modules (#119)
* Split htpy into multiple private modules

* Do not alias markupsafe imports

We only import markupsafe in private modules, so it doesn't need to be
aliased to avoid leaking the import.

* Remove prefix on non-public element helper

It is still private as it is defined and exclusively used in
private modules or inside functions.

* Remove prefix on non-public types

These are still private as they are defined and exclusively used in
private modules.

* Remove prefix on non-public rendering helpers

These are still private as they are defined and exclusively used in
private modules.

* Remove prefix on non-public attribute helpers

These are still private as they are defined and exclusively used in
private modules.

* Set __all__ to [].

---------

Co-authored-by: Andreas Pelme <andreas@pelme.se>